+++ packages/cli/src/run.js
@@ -2,7 +2,8 @@ import consola from 'consola'
 import NuxtCommand from './command'
 import * as commands from './commands'
 import setup from './setup'
-import { startSpaces, optionSpaces, maxCharsPerLine } from './formatting'
+import { loadNuxtConfig } from './utils'
+import * from './formatting'
 
 function listCommands() {
   const options = []
