+++ packages/cli/src/utils.js
@@ -62,25 +62,3 @@ export async function loadNuxtConfig(argv) {
 
   return options
 }
-
-export function indent(count, chr = ' ') {
-  return chr.repeat(count)
-}
-
-export function indentLines(string, spaces, firstLineSpaces) {
-  const lines = Array.isArray(string) ? string : string.split('\n')
-  let s = ''
-  if (lines.length) {
-    const i0 = indent(firstLineSpaces === undefined ? spaces : firstLineSpaces)
-    s = i0 + lines.shift()
-  }
-  if (lines.length) {
-    const i = indent(spaces)
-    s += '\n' + lines.map(l => i + l).join('\n')
-  }
-  return s
-}
-
-export function foldLines(string, maxCharsPerLine, spaces, firstLineSpaces) {
-  return indentLines(wrapAnsi(string, maxCharsPerLine, { trim: false }), spaces, firstLineSpaces)
-}
